Refactor sections to react-aria-components Tabs
