+++ next/components/forms/segments/AccountSections/MyApplicationsSection/MyApplicationsList.tsx
@@ -2,7 +2,7 @@ import { formsApi } from '@clients/forms'
 import { GetFormResponseDto, GetFormResponseDtoStateEnum } from '@clients/openapi-forms'
 import { useQuery } from '@tanstack/react-query'
 import MyApplicationCardsPlaceholder from 'components/forms/segments/AccountSections/MyApplicationsSection/MyApplicationCardsPlaceholder'
-import { ApplicationsListVariant } from 'components/forms/segments/AccountSections/MyApplicationsSection/MyApplicationsHeader'
+import { ApplicationsListVariant } from 'components/forms/segments/AccountSections/MyApplicationsSection/MyApplicationsSection'
 import Pagination from 'components/forms/simple-components/Pagination/Pagination'
 import useSnackbar from 'frontend/hooks/useSnackbar'
 import { useTranslation } from 'next-i18next'
